디렉터리 정리 & nfc_write_screen과 nfc_write_service 파일 구분
